Don't submit form if non-Element inputs fail validation.
In these examples, we use the basic HTML5 form validation APIs (e.g. the
`required` attribute) for validation on non-Elements inputs.

If an Element is focused and a user submit event is triggered (e.g.
through the enter key), the DOM does not automatically trigger form
validation on the outer plain inputs. This previously allowed submission
of the form with empty `required` fields.

This change catches those cases by manually triggering HTML5 form
validation on submit.